Make the `waitFor` utilities return `Future<T>` instead of `Future<T?>`.
Currently, the `condition` function has an expected return type of `FutureOr<T>?`,
which then forces the return types for all invocations to be nullable.

Instead, for methods that actually return a null, the generic instantiation
itself can use a nullable value for `T`. That way, code that always return
non-null values do not need to check for null values.