Spawn test server in a background isolate (#285)
Currently the navigation needs to use the async webdriver because the
server is running in the same isolate, and the `sync_http` package
causes the Dart isolate to wait while the request is outstanding, and
the browser will not complete the response until the navigation is
complete. If the server is in the same isolate it will not be able to
serve the content and it causes a deadlock.

Run the server in another isolate. Use isolate ports to communicate the
timing for server shutdown.

Fix some tests which relied on the navigation _not_ happening through
the sync client so that they didn't get recorded by the event listener.
Clear the events after the initial navigation to ignore non-relevant
events.